refactor:handleLogin function extracted from Login
